Move sub__depthbuffer() and sub__maptriangle() to graphics.cpp
