+++ internal/c/common.h
@@ -99,53 +99,8 @@
 #    endif
 
 #    include "error_handle.h"
+#    include "graphics.h"
 #    include "qbs.h"
-
-struct img_struct {
-    void *lock_offset;
-    int64 lock_id;
-    uint8 valid;   // 0,1 0=invalid
-    uint8 text;    // if set, surface is a text surface
-    uint8 console; // dummy surface to absorb unimplemented console functionality
-    uint16 width, height;
-    uint8 bytes_per_pixel;  // 1,2,4
-    uint8 bits_per_pixel;   // 1,2,4,8,16(text),32
-    uint32 mask;            // 1,3,0xF,0xFF,0xFFFF,0xFFFFFFFF
-    uint16 compatible_mode; // 0,1,2,7,8,9,10,11,12,13,32,256
-    uint32 color, background_color, draw_color;
-    uint32 font;               // 8,14,16,?
-    int16 top_row, bottom_row; // VIEW PRINT settings, unique (as in QB) to each "page"
-    int16 cursor_x, cursor_y;  // unique (as in QB) to each "page"
-    uint8 cursor_show, cursor_firstvalue, cursor_lastvalue;
-    union {
-        uint8 *offset;
-        uint32 *offset32;
-    };
-    uint32 flags;
-    uint32 *pal;
-    int32 transparent_color; //-1 means no color is transparent
-    uint8 alpha_disabled;
-    uint8 holding_cursor;
-    uint8 print_mode;
-    // BEGIN apm ('active page migration')
-    // everything between apm points is migrated during active page changes
-    // note: apm data is only relevant to graphics modes
-    uint8 apm_p1;
-    int32 view_x1, view_y1, view_x2, view_y2;
-    int32 view_offset_x, view_offset_y;
-    float x, y;
-    uint8 clipping_or_scaling;
-    float scaling_x, scaling_y, scaling_offset_x, scaling_offset_y;
-    float window_x1, window_y1, window_x2, window_y2;
-    double draw_ta;
-    double draw_scale;
-    uint8 apm_p2;
-    // END apm
-};
-// img_struct flags
-#    define IMG_FREEPAL 1 // free palette data before freeing image
-#    define IMG_SCREEN 2  // img is linked to other screen pages
-#    define IMG_FREEMEM 4 // if set, it means memory must be freed
 
 // QB64 internal variable type flags (internally referenced by some C functions)
 #    define ISSTRING 1073741824
